🚧 mturk ad page support
explicitly include routes to /ad and /mturk.js for amazon turk.
the ad page looks at assignmentId to see if its a preview or not.
accepting goes to the actual task. but we might need a consent page.
a consent page or at least a wrapper/redirect might be useful to
capture workerId in addition to assignmentId. This will need more work
on the database.

Currently not concerned about hitId <-> task lookup. but that could
also be stored somewhere.