🐛mturk.js error obstructed launching task from /ad
I tried to add tests, but how to mock open.window is not immediately obvious
I think the issue was the function popup defining a variable popup.

Also now catch the bad response if the server fails to respond to /anchor/$task
This is useful for testing the JS without a server. but will maybe
allow something that should break to continue